Moved fetch requests into shared lib
